Ignore Kubernetes/OpenShift components that have deployByDefault=false
Since deployByDefault=false components are not intended to be deployed
when the DevWorkspace is started, remove restrictions/checks on these
components -- i.e. URI is allowed, and we don't check RBAC permissions.

These components are meant to be deployed later, e.g. via an apply
command, and so verifying RBAC permissions on them is up to the tool
used to apply them (e.g. by using the user's token to apply the
resources).

Signed-off-by: Angel Misevski <amisevsk@redhat.com>