Avoid waiting 5 seconds for image pull secret on every k8s start
Recent versions of Kubernetes do not automatically mount image pull
secrets to serviceaccounts, meaning that most workspace starts idle for
5 seconds waiting for something that is not going to happen. Instead we
scope the check for waiting for image pull secrets to only be performed
when running on OpenShift.

On OpenShift, there is a race between the controller and the cluster
API, where sometimes DWO creates a deployment that is immediately
updated with image pull secrets, slowing workspace start due to the
workspace pod needing to terminate.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>