Don't set controllerref on shared dw ServiceAccount
Since objects can only have one controller ownerref, don't attempt to
set an controller for the shared workspace serviceAccount when it is
configured. Instead set regular ownerrefs only to ensure serviceaccount
is cleaned up when workspaces are deleted.

This requires updating how the objects are compared and update in order
to ensure we can append additional ownerrefs when others are present.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>