Support defining strategic patch directives in override attributes
Add support for Kubernetes strategic merge directives, allowing more
control over pod/container overrides. For example, the attribute

    pod-overrides:
      spec:
        securityContext:
          runAsUser: 1001
          $patch: replace

can be used to overwrite the default securityContext for a pod -- if the
default pod SecurityContext is configured to be

    podSecurityContext:
      fsGroup: 2000
      runAsGroup: 3000
      runAsUser: 1000

using $patch: replace will result in the pod SecurityContext

    podSecurityContext:
      runAsUser: 1001

instead of

    podSecurityContext:
      fsGroup: 2000
      runAsGroup: 3000
      runAsUser: 1001

To achieve this, we use the raw patch as defined in the overrides
attribute rather than deserializing and reserializing it.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>