feat: support PVC expansion based on DWOC
PVC expansion is an experimental feature and must be enabled
by setting `enableExperimentalFeatures` to `true` in the DWOC.

PVC expansion occurs when DWO sees that a PVC that was provisioned
under the `common` (or `per-user`) or `per-workspace` storage-type is smaller
than the size specified in the DWOC's `workspace.defaultStorageSize`.

This feature only works when the Kubernetes StorageClass that provisions
the PVC has the allowVolumeExpansion field set to `true`. Additionally, the
volume must also support expansion.

See https://kubernetes.io/docs/concepts/storage/persistent-volumes/#expanding-persistent-volumes-claims
for a list of volumes that support expansion.

Signed-off-by: Andrew Obuchowicz <aobuchow@redhat.com>