Update component for commands that target merged contributions
When a container contribution is merged, the original component is no
longer included in the devfile. This breaks any commands that reference
that component, potentially resulting in an invalid DevWorkspace.

Instead, update any commands that target the contribution component to
instead target the component it was merged into.

Components that specify they are merge contributions and are also
targeted by an apply command result in an error, as merging the
component makes the apply command invalid.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>