Remove timing annotations functionality from operator
Remove experimental feature that adds annotations workspace startup
timing. This feature is obsoleted by workspace conditions, which
can be used to gain the same information (and are currently used in
reporting metrics)

The approach of adding timing information in annotations on the
DevWorkspace object also led to quite a few conflicts in the API,
resulting in misleading messages being logged by the DevWorkspace
Operator.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>