Add logs when waiting on workspace mainUrl
Avoid exiting the reconcile loop without logging anything in order to
make logs clearer (so that every "reconciling workspace..." log has a
corresponding "waiting for <x>" log).

This resolves an issue where a workspace stuck in "waiting for editor to
be ready" would log no information for reconciles, making it unclear if
the workspace ever reached a running state.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>