Rework project-clone checkout refernce to not rely on go-git
To avoid issues with self-signed certificates on git repositories, avoid
use of go-git (which would require configuring additional CA bundles)
and instead delegate to the git binary, which respects additional
bundles automatically.

The main cause for failure is go-git attempting to list references in a
remote, which can fail due to untrusted certificates. Instead, we rely
on the git binary to determine if a checkoutFrom.revision is a
local/remote branch, tag, or hash.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>