Add DevWorkspace for semi-automatically testing project-clone container
Add devworkspace project-clone/test/project-clone-test.devworkspace.yaml
to allow for semi-automatic testing of the project-clone container. This
workspace is set up to run a bash script (stored in the .command field)
on start that verifies that projects have been set up as expected.

If projects were not set up correctly, the workspace enters a failed
state and logs for the workspace container can be checked for the cause
of the failure.

The workspace makes use of devfile variables to enable configuration of
the key elements being tested (main repo, fork repo, branch, tag,
revision)

Signed-off-by: Angel Misevski <amisevsk@redhat.com>