Do not use default CPU limit on project clone if possible
If the workspace's namespace does not define any limitRanges that would
prevent creation of a pod that does not specify a CPU limit, do not set
a CPU limit for the project clone init container by default.

If the operator's configuration specifies a non-default CPU limit for
project-clone, that value is always used.

Currently, the controller only checks whether limitRanges are present,
so a limit range that only applies to PVCs will still result in a
default CPU limit being used.

Co-authored-by: Anatolii Bazko <abazko@redhat.com>
Signed-off-by: Angel Misevski <amisevsk@redhat.com>