Update contribution merging to use default resources as base
Apply default resources to any merge target when flattening a
devworkspace in order to correctly apply default resource requirements.
Otherwise, merging into a component that does not specify e.g. a memory
limit will result in the default memory limit being used for the merged
component, which is incorrect.

Since merge contributions do not add resources when the merge target
does not define a field (e.g. CPU limit), we need to apply the default
resources to the component earlier in order to correctly calculate the
correct total.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>