Reimplement project-clone subPath handling to conform to devfile API
The subDir field on starterProjects is intended to take one directory
out of a starterProject rather than perform a sparse checkout.

The previous subDir handling as been reworked into a regular
sparse-checkout by renaming the attribute 'sparseCheckout'. The 'subDir'
attribute now configures the project-clone container to

1. Clone the project to a temporary directory, set up remotes, check out
   a branch, etc.
2. Copy _the subdirectory specified by subDir_ from that project into
   $PROJECTS_ROOT

This results in a folder within $PROJECTS_ROOT that is not a git
repository and only contains the contents of that subfolder.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>