Allow cluster-detected proxy settings to be ignored in OpenShift
Change httpProxy, httpsProxy, and noProxy fields in DWOC's proxyConfig
field to be string pointers rather than strings. This allows
distinguishing between "not set" and "set to the empty string" in the
DevWorkspace Operator config, which is necessary to allow the admin to
"disable" automatically-detected proxy settings by setting them to the
empty string.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>