Wrap git credential helper to ignore 'store' and 'erase' operations
Since the credentials file is mounted read-only to the filesystem, we
need to configure a git credential helper that ignores 'store' and
'erase' operations in order to avoid printing error messages around
getting a lock on the credentials file.

In order to do this, we wrap the existing 'store' credential helper to
effectively be read-only.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>