fix: don't abort PVC size calculation early if volume size not defined
Fix #1162

The following rules are supposed to be used when computing the per-workspace PVC size:
1. If all volumes in a devworkspace specify their size, the computed PVC size will be used.
2. If all volumes in a devworkspace either specify their size or are ephemeral, the computed PVC size will be used.
3. If at least one volume in a devworkspace specifies its size, and the computed PVC size is greater than the default per-workspace PVC size, the computed PVC size will be used.

Prior to this commit, rule 3 was not being respected in cases where a volume did not define its size, but later volumes (in the volume array)
did define their sizes.

This commit prevents aborting the PVC size calculation too early, and modifies the test case that was relevant to rule 3 to ensure this case is accounted for.

Signed-off-by: Andrew Obuchowicz <aobuchow@redhat.com>