Handle failures in routing reconciler in recoverable way
When a DevWorkspaceRouting fails to provision objects, it enters the
'Failed' phase. In order to prevent repeated unnecessary reconciles, the
controller ignores DWRs that are already Failed, which leads to the
'Failed' phase being permanent.

To avoid this, we instead set a Stopped phase when the parent workspace
is stopped. This will clear the Failed phase and any message that was
set, so the DevWorkspace controller will need to propagate it into the
DevWorkspace object to allow it to be shown.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>