[project-clone] Initialize and update submodules when present in project
If a cloned-and-checked-out project has a .gitmodules file in the root
of the repo, run 'git submodule update --init --recursive' inside the
project to clone all submodules.

Failures in setting up submodules result only in a warning log and do
not prevent the project from being 'set up', as the user can later
initialize submodules manually.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>