Avoid syncing no-op status changes for DevWorkspaces
Before updating a workspace's status, check that the update actually
updates the status. If the current status and on-cluster status are
identical, do not make the update request.

This change is useful for improving processing of workspaces in bulk,
either during operator start up or on the regularly-timed sync interval
(default: ~10h). When re-reconciling a workspace that has seen no
changes, we don't actually have to make any requests to the API server.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>