fix: temporarily pin goimports in GH Actions to v0.24.0
Part of #1314

The latest version of goimports (v0.25.0) requires go >= 1.22.0.

Since we are about to do the upstream release of DWO 0.31.0, we are
temporarily pinning goimports to the last working version (v0.24.0).

After the 0.31.0 release of DWO, we can safely upgrade the entire
repo to go 1.22.0, and use the latest version of goimports again.

Signed-off-by: Andrew Obuchowicz <aobuchow@redhat.com>