fix: conditionally add SSH agent postStart event
Only add the SSH agent initialization postStart event if
an SSH key with a passphrase is being used & experimental features are
enabled.

We don't use the config package's ExperimentalFeaturesEnabled function so that
the SSH agent initialization postStart event injection can be enabled from
an external DWOC, or the global DWOC if no external DWOC is used.

fix #1340

Signed-off-by: Andrew Obuchowicz <aobuchow@redhat.com>