feat(23570): Add controller for workspace backup
A new backup controller orchestrates a backup process for workspace PVC.
A new configuration option is added to DevWorkspaceOperatorConfig that
enables running regular cronjob that is responsible for backup
mechanism. The job executes following steps:
- Find a workspaces
- Finds out that workspace has been recently stopped
- Detect a workspace PVC
- Execute a job in the same namespace that does the backup

The last step is currently not fully implemented as it requires running
a buildah inside the container and it will be delivered as a separate
feature.

Issue: https://github.com/eclipse-che/che/issues/23570

Signed-off-by: Ales Raszka <araszka@redhat.com>